requirements.txt
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.0
seaborn==0.12.2
streamlit==1.30.0
pickle-mixin==1.0.2
pillow==9.4.0
scikit-learn==1.2.1